Explicitly setup heap, fixes brahma.
Since brahma has its stack (unless it is changed) in ARM9 memory, the
default logic for setting up the heap breaks, since it assumes the heap
is in lower memory than the stack. This commit uses a hook used by
newlib to setup the heap at any random location in memory. This commit
just allocates a 32MB heap, although it could certainly be larger.